% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{as_matrix}
\alias{as_matrix}
\title{Convert cor_df to original matrix.}
\usage{
as_matrix(x)
}
\arguments{
\item{x}{cor_df. See \code{\link{correlate}}.}
}
\value{
Correlation matrix
}
\description{
Convert a correlation data frame to original matrix format.
}
\examples{
x <- correlate(mtcars)
as_matrix(x)
}

