% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_x.R
\name{cor_x}
\alias{cor_frame}
\alias{cor_matrix}
\alias{cor_x}
\title{Obtain correlations and corresponding sample sizes.}
\usage{
cor_matrix(x, y = NULL, use = "pairwise.complete.obs", method = "pearson")

cor_frame(x, y = NULL, use = "pairwise.complete.obs", method = "pearson")
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with
    compatible dimensions to \code{x}.   The default is equivalent to
    \code{y = x} (but more efficient).}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}
}
\description{
These are extensions of stats::cor(). They use cor() to calculate
correlations, and return the results in a data frame or matrix-list structure
with corresponding sample sizes.
}
\section{Return structures}{


Currently corrr returns correlations and sample sizes in two structures:

\describe{
 \item{\code{cor_matrix}}{return a list of two matrices: \code{r}, which
 contains the correlations; and \code{n}, which contains the corresponding
 sample sizes.}
 
 \item{\code{cor_frame}}{return a dplyr::data_frame with three columns:
 \code{vars}, the column names (variables) being correlated pasted together
 with \code{<>}; \code{r}, which contains the correlations; and \code{n}, which
 contains the corresponding sample sizes. Unlike \code{cor_matrix},
 \code{cor_frame} only returns each correlation once (rather than in a
 mirrored matrix).}
 
}
}

