% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plus.R
\name{corr_plus}
\alias{corr_plus}
\title{`corr_plus` function coefficients calculates the relationship between
two variables as well as the magnitude of this relationship.}
\usage{
corr_plus(x, y)
}
\arguments{
\item{x}{a vector or an array of a random variable}

\item{y}{a vector or an array of a random variable}
}
\value{
a single value correlation between two variables
}
\description{
`corr_plus` function coefficients calculates the relationship between
two variables as well as the magnitude of this relationship.
}
