% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{cross_cor}
\alias{cross_cor}
\title{Cross correlate.}
\usage{
cross_cor(x, ..., in_rows = TRUE)
}
\arguments{
\item{x}{A \code{\link{correlate}} correlation matrix.}

\item{...}{Vector of/comma separated character strings of variable names.}

\item{in_rows}{Boolean to keep the named variables in the rows, or
columns otherwise. Default = TRUE.}
}
\description{
Convenience function to cross correlate one set of variables with all others
from a correlation matrix. This function will take a \code{\link{correlate}}
correlation matrix, and keep the named variables in the rows (or columns),
and drop them from the columns (or rows).
}

