% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{cross_section}
\alias{cross_section}
\title{Cross section of the correlation matrix.}
\usage{
cross_section(x, ..., keep_in_row = TRUE)
}
\arguments{
\item{x}{A \code{\link{correlate}} correlation matrix.}

\item{...}{Vector of/comma separated character strings of variable names.}

\item{keep_in_row}{Boolean to keep the named variables in the rows, or
columns otherwise. Default = TRUE.}
}
\description{
Convenience function to create a cross section of the matrix by selecting a 
set of variables to be kept in one axis (rows or columns), and dropped from
the other.
}

