% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{dice}
\alias{dice}
\title{Returns a correlation table with the selected fields only}
\usage{
dice(x, ...)
}
\arguments{
\item{x}{A correlation table, class cor_df}

\item{...}{A list of variables in the correlation table}
}
\description{
Returns a correlation table with the selected fields only
}
\examples{

dice(correlate(mtcars), mpg, wt, am)
}
