% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{fashion}
\alias{fashion}
\title{Fashion a correlation data frame for printing.}
\usage{
fashion(x, decimals, na_print)
}
\arguments{
\item{x}{Scalar, vector, matrix or data frame.}

\item{decimals}{Number of decimal places to display for numbers.}

\item{na_print}{Character string indicating NA values in printed output}
}
\value{
noquote. Also a data frame if a x is a matrix or data frame.
}
\description{
For the purpose of printing, convert a correlation data frame into a noquote 
matrix with the correlations cleanly formatted (leading zeros removed; spaced
for signs) and the diagonal (or any NA) left blank.
}

