% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{first_col}
\alias{first_col}
\title{Add a first column to a data.frame}
\usage{
first_col(df, ..., var = "rowname")
}
\arguments{
\item{...}{Values to go into the column}

\item{var}{Label for the column. Default is "rowname"}

\item{x}{Data frame}
}
\description{
Add a first column to a data.frame. This is most commonly used toappend a
rowname column to create a cor_df.
}
\examples{
first_col(mtcars, 1:nrow(mtcars))
}

