% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_x.R
\name{na_x}
\alias{na_lower}
\alias{na_upper}
\alias{na_x}
\title{Set portion to missing}
\usage{
na_upper(x)

na_lower(x)
}
\arguments{
\item{x}{cor_df}
}
\description{
Convert a portion of a correlation data frame (cor_df) to missing.
}
\section{Portions to convert to missing}{


Currently, corrr supports the setting of the upper and lower triangles to
missing.

\describe{
  \item{\code{na_upper}}{convert the upper matrix to missing}
  
  \item{\code{na_lower}}{convert the lower triangle to missing}
  
}
}

