% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pair_n}
\alias{pair_n}
\title{Number of pairwise complete cases.}
\usage{
pair_n(x, y = NULL)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{y}{\code{NULL} (default) or a vector, matrix or data frame with
    compatible dimensions to \code{x}.   The default is equivalent to
    \code{y = x} (but more efficient).}
}
\value{
Matrix of pairwise sample sizes (number of complete cases).
}
\description{
Compute the number of complete cases in a pairwise fashion for \code{x} (and
\code{y}).
}
\examples{
pair_n(mtcars)
}

