% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{rarrange}
\alias{rarrange}
\title{Re-arrange a correlation matrix}
\usage{
rarrange(x, method = "PC", absolute = TRUE)
}
\arguments{
\item{x}{A \code{\link{correlate}} correlation matrix.}

\item{method}{String specifying the arrangement (clustering) method.
Clustering is achieved via \code{\link[seriation]{seriate}}, which can be
consulted for a complete list of clustering methods. Default = "PCA".}

\item{absolute}{Boolean whether absolute values for the correlations should
be used when clustering.}
}
\description{
Re-arrange a correlation matrix to group highly correlated variables closer
together.
}

