% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retract.R
\name{retract}
\alias{retract}
\title{Creates a data frame from a stretched correlation table}
\usage{
retract(.data, x, y, val)
}
\arguments{
\item{.data}{A data.frame or tibble containing at least three variables: x, y and the value}

\item{x}{The name of the column to use from .data as x}

\item{y}{The name of the column to use from .data as y}

\item{val}{The name of the column to use from .data to use as the value}
}
\description{
\code{retract} does the opposite of what \code{stretch} does
}
\examples{
x <- correlate(mtcars)
xs <- stretch(x)
retract(xs)
}
