% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{rselect}
\alias{rselect}
\title{Select columns, exclude same or other rows.}
\usage{
rselect(x, ..., rows = TRUE)
}
\arguments{
\item{x}{A \code{\link{correlate}} correlation matrix.}

\item{...}{Comma separated list of unquoted expressions. You can treat
variable names like they are positions. Use positive values to select
variables; use negative values to drop variables.}

\item{rows}{Boolean. Whether to keep selected variables in the rows, or all other
variables otherwise.}
}
\description{
Convenience function to select a set of variables from a correlation matrix
to keep as the columns, and exclude these or all other variables from the rows. This
function will take a \code{\link{correlate}} correlation matrix, and
expression(s) suited for dplyr::select(). The selected variables will remain
in the columns, and these, or all other variables, will be excluded from the
rows based on `\code{same}. For a complete list of methods for using this
function, see \code{\link[dplyr]{select}}.
}
\examples{
mtcars \%>\% correlate() \%>\% rselect(mpg, cyl)
mtcars \%>\% correlate() \%>\% rselect(-disp, - mpg)
iris[, 1:4] \%>\% correlate() \%>\% rselect(-matches("Sepal"))

}

