% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_plus.R
\name{std_plus}
\alias{std_plus}
\title{`std_plus` function calculates calculates how close the data
points are to the mean and the variation of the data points.}
\usage{
std_plus(x)
}
\arguments{
\item{x}{a vector or an array of a random variable}
}
\value{
the value of standard deviation of the input data
}
\description{
`std_plus` function calculates calculates how close the data
points are to the mean and the variation of the data points.
}
