% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{subspace}
\alias{subcols}
\alias{subcols_}
\alias{subrows}
\alias{subrows_}
\alias{subspace}
\title{Select a subspace of variables by name.}
\usage{
subrows(.r_mat, ...)

subrows_(.r_mat, ..., .dots)

subcols(.r_mat, ...)

subcols_(.r_mat, ..., .dots)
}
\arguments{
\item{.r_mat}{A correaltion matrix with row and column names.}

\item{...}{Comma separated names of the rows/columns to keep.}
}
\description{
Select a subspace of only the variables you mention by name in rows or cols.
}
\section{Subspace methods}{


\describe{
  \item{\code{subrows}}{Return all rows whose names match those provided.}
  
  \item{\code{subcols}}{Return all columns whose names match those provided.}
  
  \item{\code{subrows_} and \code{subrows_}}{Programattic versions of above
  that accept string characters as row/column names.}
}
}

