% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{xselect}
\alias{xselect}
\title{Select variables as columns, remove as rows.}
\usage{
xselect(x, ...)
}
\arguments{
\item{x}{A \code{\link{correlate}} correlation matrix.}

\item{...}{Comma separated list of unquoted expressions. You can treat
variable names like they are positions. Use positive values to select
variables; use negative values to drop variables.}
}
\description{
Convenience function to select a set of variables from a correlation matrix
to keep as the columns, but remove these same variables from the rows. This
function will take a \code{\link{correlate}} correlation matrix, and
expression(s) suited for dplyr::select(). The selected variables will remain
in the columns, but be removed from the rows. For a complete list of methods
for using this function, see \code{\link[dplyr]{select}}.
}
\examples{
mtcars \%>\% correlate() \%>\% xselect(mpg, cyl)
mtcars \%>\% correlate() \%>\% xselect(-disp, - mpg)
iris[, 1:4] \%>\% correlate() \%>\% xselect(-matches("Sepal"))

}

