% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{as_matrix}
\alias{as_matrix}
\title{Convert cor_df to original matrix.}
\usage{
as_matrix(x)
}
\arguments{
\item{x}{A \code{\link{correlate}} correlation matrix.}
}
\description{
Convert a \code{\link{correlate}} correlation matrix from data frame to
original matrix format.
}

