% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{as_matrix}
\alias{as_matrix}
\title{Convert a correlation data frame to matrix format}
\usage{
as_matrix(x, diagonal = 1)
}
\arguments{
\item{x}{A correlation data frame. See \code{\link{correlate}} or \code{\link{as_cordf}}.}

\item{diagonal}{Value (typically numeric or NA) to set the diagonal to.}
}
\value{
Correlation matrix
}
\description{
Convert a correlation data frame to original matrix format.
}
\examples{
x <- correlate(mtcars)
as_matrix(x)
}

