% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_x.R
\name{cor_x}
\alias{cor_frame}
\alias{cor_matrix}
\alias{cor_x}
\title{Obtain correlations and corresponding sample sizes.}
\usage{
cor_matrix(x, use = "pairwise.complete.obs", method = c("pearson",
  "kendall", "spearman"))

cor_frame(x, use = "pairwise.complete.obs", method = c("pearson", "kendall",
  "spearman"))
}
\arguments{
\item{x}{A data frame of columns to correlate.}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}
}
\description{
These are methods for calculating the correlations (and corresponding sample
sizes) between columns of a data frame.
}
\section{Return structures}{


Currently corrr returns correlations and sample sizes in two structures:

\describe{
 \item{\code{cor_matrix}}{return a list of two matrices: \code{r}, which
 contains the correlations; and \code{n}, which contains the corresponding
 sample sizes.}
 
 \item{\code{cor_frame}}{return a dplyr::data_frame with three columns:
 \code{vars}, the column names (variables) being correlated pasted together
 with \code{<>}; \code{r}, which contains the correlations; and \code{n}, which
 contains the corresponding sample sizes. Unlike \code{cor_matrix},
 \code{cor_frame} only returns each correlation once (rather than in a
 mirrored matrix).}
 
}
}

