% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plus.R
\name{corr_plus}
\alias{corr_plus}
\title{Correlation}
\usage{
corr_plus(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
the value of standard deviation of the input data
}
\description{
`corr_plus` function calculates the Pearson correlation coefficient between
two variables using complete case analysis that excludes all observations
that have NA for at least one variable.
}
\examples{
corr_plus(c(1,2,3,4,5), c(1,2,3,NA,5))
}
