% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrr.R
\docType{package}
\name{corrr-package}
\alias{corrr}
\alias{corrr-package}
\title{corrr: Correlations in R}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

A tool for exploring correlations.
    It makes it possible to easily perform routine tasks when
    exploring correlation matrices such as ignoring the diagonal,
    focusing on the correlations of certain variables against others,
    or rearranging and visualising the matrix in terms of the
    strength of the correlations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/corrr}
  \item Report bugs at \url{https://github.com/tidymodels/corrr/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgararuiz@gmail.com}

Authors:
\itemize{
  \item Simon Jackson
  \item Jorge Cimentada
}

}
\keyword{internal}
