% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrR-package.R
\docType{package}
\name{CorrR-package}
\alias{CorrR}
\alias{CorrR-package}
\title{CorrR: A package for calculating correlation coefficients and covariance matrix}
\description{
A R package that calculates correlation coefficients and covariance matrix of a given data with missing values. 
             The package contains std_plus, corr_plus and cov_mx.
}
\author{
\strong{Maintainer}: Mengda (Albert) Yu \email{albertlife4913@gmail.com}

}
\keyword{internal}
