% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mx.R
\name{cov_mx}
\alias{cov_mx}
\title{`cov_mx` function calculates the covariance matrix of the two
 variables and automatically deals with the missing value}
\usage{
cov_mx(x)
}
\arguments{
\item{x}{a matrix that contains one or more than one random variable}
}
\value{
matrix that contains the covariance between random variables in the input matrix
}
\description{
`cov_mx` function calculates the covariance matrix of the two
 variables and automatically deals with the missing value
}
