% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{fashion}
\alias{fashion}
\title{Fashion a correlation data frame for printing.}
\usage{
fashion(x, digits)
}
\arguments{
\item{x}{cor_df. See \code{\link{correlate}}.}

\item{digits}{Number of decimal places to display.}
}
\value{
Formatted noquote matrix.
}
\description{
For the purpose of printing, convert a correlation data frame into a noquote 
matrix with the correlations cleanly formatted (leading zeros removed; spaced
for signs) and the diagonal (or any NA) left blank.
}

