% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{first_col}
\alias{first_col}
\title{Add a first column to a data.frame}
\usage{
first_col(df, ..., var = "term")
}
\arguments{
\item{df}{Data frame}

\item{...}{Values to go into the column}

\item{var}{Label for the column, with the default "term"}
}
\description{
Add a first column to a data.frame. This is most commonly used to append a
term column to create a cor_df.
}
\examples{
first_col(mtcars, 1:nrow(mtcars))
}
