% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{focus}
\alias{focus}
\title{Focus on section of a correlation data frame.}
\usage{
focus(x, ..., mirror = FALSE)
}
\arguments{
\item{x}{cor_df. See \code{\link{correlate}}.}

\item{...}{Comma separated list of unquoted expressions. You can treat
variable names like they are positions. Use positive values to select
variables; use negative values to drop variables.}

\item{mirror}{Boolean. Whether to mirror the selected columns in the rows or
not.}
}
\description{
Convenience function to select a set of variables from a correlation matrix
to keep as the columns, and exclude these or all other variables from the rows. This
function will take a \code{\link{correlate}} correlation matrix, and
expression(s) suited for dplyr::select(). The selected variables will remain
in the columns, and these, or all other variables, will be excluded from the
rows based on `\code{same}. For a complete list of methods for using this
function, see \code{\link[dplyr]{select}}.
}
\examples{
mtcars \%>\% correlate() \%>\% focus(mpg, cyl)
mtcars \%>\% correlate() \%>\% focus(-disp, - mpg, mirror = TRUE)
iris[, 1:4] \%>\% correlate() \%>\% focus(-matches("Sepal"))

}

