% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rplot}
\alias{rplot}
\title{Plot correlation matrix.}
\usage{
rplot(x)
}
\arguments{
\item{x}{A \code{\link{correlate}} correlation matrix.}
}
\description{
Plot a \code{\link{correlate}} correlation matrix.
}

