% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rplot}
\alias{rplot}
\title{Plot correlation matrix.}
\usage{
rplot(x, shape)
}
\arguments{
\item{x}{cor_df. See \code{\link{correlate}}.}

\item{shape}{\code{\link{geom_point}} aesthetic.}
}
\description{
Plot a \code{\link{correlate}} correlation matrix.
}

