% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{shave}
\alias{shave}
\title{Shave off upper/lower triangle.}
\usage{
shave(x, upper = TRUE)
}
\arguments{
\item{x}{cor_df. See \code{\link{correlate}}.}

\item{upper}{Boolean. If TRUE, set upper triangle to NA; lower triangle if
FALSE.}
}
\value{
cor_df. See \code{\link{correlate}}.
}
\description{
Convert the upper or lower triangle of a correlation data frame (cor_df) to
missing values.
}
\examples{
x <- correlate(mtcars)
shave(x)  # Default; shave upper triangle
shave(x, upper = FALSE)  # shave lower triangle
}
