% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{stretch}
\alias{stretch}
\title{Stretch correlation data frame into long format.}
\usage{
stretch(x, ..., na_omit = FALSE)
}
\arguments{
\item{x}{cor_df. See \code{\link{correlate}}.}

\item{...}{Comma separated list of unquoted expressions. You can treat
variable names like they are positions. Use positive values to select
variables; use negative values to drop variables.}

\item{na_omit}{Boolean. Whether rows with an NA correlation (originally the
matrix diagonal) should be dropped? Will automatically be set to TRUE if
mirror is FALSE.}
}
\description{
\code{stretch} is an extension of tidyr::gather() to be applied to a
\code{\link{correlate}} correlation matrix. It will gather the selected
variables into a long-format data frame. The rowname column is handled
automatically. Variable selection is achieved with \code{\link{focus}},
which you can see for more detail on how to select columns.
}

